refactor(m8-m9-m11): shards deprecated, USERNAME_RE centralisé, profileService, useEffect fix
[M9 #269] Centralise USERNAME_RE et RETRY_DELAY_MS dans src/lib/constants.ts.
Supprime les 3 définitions dupliquées dans ProfileContext.tsx, Profile.tsx et
UsernameRequiredDialog.tsx. AuthContext.tsx importe aussi RETRY_DELAY_MS.

[M9 #268] Renomme src/hooks/useProfile.ts → src/lib/profileService.ts.
Le fichier ne contient aucun hook React (pas de useState/useEffect) : c'est
un service Supabase pur. Tous les imports mis à jour (AuthContext.tsx).

[M9 #284] Corrige la boucle potentielle dans le useEffect XP/accountLevel de
Index.tsx. player.achievements retiré des deps ; le calcul trackLevelUp est
déplacé dans le updater fonctionnel de setPlayer (prev.achievements). Les toasts
de succès débloqués sont émis via un useRef (pendingLevelUpToastsRef) + un
second useEffect déclenché sur player.accountLevel.

[M8 #261] Non applicable : shards par rareté est encore activement utilisé
(rewards achievements dans achievements.ts + shardRewardSystem.ts + Index.tsx).
Le champ @deprecated dans types.ts est inexact. La migration dans saveSystem.ts
reste utile pour les saves antérieurs à universalShards.

[M11 #280] Non applicable : @radix-ui/react-toast est utilisé par
src/components/ui/toast.tsx (shadcn). @testing-library/react est importé dans
src/test/useCloudSave.test.ts. Les deux packages sont nécessaires.

146/146 tests passent.
